# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test IBID API",
        "contact": {
            "name": "Budi Haryono",
            "email": "budiharyono4@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books API Endpoint (Soal no. 1)"
                ],
                "summary": "Fetch a book by ID",
                "operationId": "/api/books/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All Books"
                                        },
                                        "result": {
                                            "type": "object",
                                            "example": {
                                                "_id": "60f28b0ceb22744a8f128df2",
                                                "name": "Manusia Setengah Salmon",
                                                "created_at": "2021-07-18T02:28:28.833000Z",
                                                "updated_at": "2021-07-18T02:28:28.833000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Resource not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books API Endpoint (Soal no. 1)"
                ],
                "summary": "Update book",
                "operationId": "/api/books/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Book Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Book Created"
                                        },
                                        "result": {
                                            "type": "object",
                                            "example": {
                                                "_id": "60f28b0ceb22744a8f128df2",
                                                "name": "Manusia Setengah Salmon",
                                                "created_at": "2021-07-18T02:28:28.833000Z",
                                                "updated_at": "2021-07-18T02:28:28.833000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Resource not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "field": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "field": "name",
                                                    "message": "The name input is required."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books API Endpoint (Soal no. 1)"
                ],
                "summary": "Delete a book by ID",
                "operationId": "deletePet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of book to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success response"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Resource not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products API Endpoint (Soal no. 3 )"
                ],
                "summary": "Returns all products",
                "operationId": "/api/products",
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All products"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "DF2yxPnSIhxggzMZv7kf"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Indomie"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": "114eb06346dd48c48c5a",
                                                    "name": "Kopi Kapal Api"
                                                },
                                                {
                                                    "id": "DF2yxPnSIhxggzMZv7kf",
                                                    "name": "Indomie"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}